(FEAT) : 영화 상세 조회 로직 분리
